+++ Packages/Accelib/Runtime/Module.SceneManagement/Utility/SceneChanger.cs
@@ -10,9 +10,9 @@ namespace Accelib.Module.SceneManagement.Utility
     {
         [Title("옵션")]
         [SerializeField] private LoadSceneMode loadMode = LoadSceneMode.Single;
-        
+
         [ShowIf("loadMode", LoadSceneMode.Single)]
-        [SerializeField] private AssetReference sceneAsset;
+        [SerializeField] private SceneRef sceneAsset;
         [ShowIf("loadMode", LoadSceneMode.Additive)]
         [SerializeField] private AssetReference[] sceneAssets;